Migrate refit function
